Split bucket interface
