+++ radix-engine-interface/src/blueprints/resource/bucket.rs
@@ -1,17 +1,13 @@
 use crate::data::scrypto::model::Own;
-use crate::data::scrypto::model::*;
 use crate::data::scrypto::ScryptoCustomTypeKind;
 use crate::data::scrypto::ScryptoCustomValueKind;
 use crate::math::*;
 use crate::*;
 use radix_engine_common::types::*;
-use sbor::rust::collections::BTreeSet;
 use sbor::rust::fmt::Debug;
 use sbor::*;
 
 use super::Proof;
-
-pub const BUCKET_BLUEPRINT: &str = "Bucket";
 
 pub const BUCKET_BURN_IDENT: &str = "burn_bucket";
 
@@ -30,7 +26,7 @@ impl Clone for BucketBurnInput {
     }
 }
 
-pub const BUCKET_DROP_EMPTY_IDENT: &str = "Bucket_drop_empty";
+pub const BUCKET_DROP_EMPTY_IDENT: &str = "drop_empty";
 
 #[derive(Debug, Eq, PartialEq, ScryptoSbor)]
 pub struct BucketDropEmptyInput {
@@ -39,7 +35,7 @@ pub struct BucketDropEmptyInput {
 
 pub type BucketDropEmptyOutput = ();
 
-pub const BUCKET_TAKE_IDENT: &str = "Bucket_take";
+pub const BUCKET_TAKE_IDENT: &str = "take";
 
 #[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
 pub struct BucketTakeInput {
@@ -48,7 +44,7 @@ pub struct BucketTakeInput {
 
 pub type BucketTakeOutput = Bucket;
 
-pub const BUCKET_PUT_IDENT: &str = "Bucket_put";
+pub const BUCKET_PUT_IDENT: &str = "put";
 
 #[derive(Debug, Eq, PartialEq, ScryptoSbor)]
 pub struct BucketPutInput {
@@ -65,78 +61,26 @@ impl Clone for BucketPutInput {
     }
 }
 
-pub const BUCKET_TAKE_NON_FUNGIBLES_IDENT: &str = "Bucket_take_non_fungibles";
-
-#[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
-pub struct BucketTakeNonFungiblesInput {
-    pub ids: BTreeSet<NonFungibleLocalId>,
-}
-
-pub type BucketTakeNonFungiblesOutput = Bucket;
-
-pub const BUCKET_GET_NON_FUNGIBLE_LOCAL_IDS_IDENT: &str = "Bucket_get_non_fungible_local_ids";
-
-#[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
-pub struct BucketGetNonFungibleLocalIdsInput {}
-
-pub type BucketGetNonFungibleLocalIdsOutput = BTreeSet<NonFungibleLocalId>;
-
-pub const BUCKET_GET_AMOUNT_IDENT: &str = "Bucket_get_amount";
+pub const BUCKET_GET_AMOUNT_IDENT: &str = "get_amount";
 
 #[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
 pub struct BucketGetAmountInput {}
 
 pub type BucketGetAmountOutput = Decimal;
 
-pub const BUCKET_GET_RESOURCE_ADDRESS_IDENT: &str = "Bucket_get_resource_address";
+pub const BUCKET_GET_RESOURCE_ADDRESS_IDENT: &str = "get_resource_address";
 
 #[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
 pub struct BucketGetResourceAddressInput {}
 
 pub type BucketGetResourceAddressOutput = ResourceAddress;
 
-pub const BUCKET_CREATE_PROOF_IDENT: &str = "Bucket_create_proof";
+pub const BUCKET_CREATE_PROOF_IDENT: &str = "create_proof";
 
 #[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
 pub struct BucketCreateProofInput {}
 
 pub type BucketCreateProofOutput = Proof;
-
-pub const BUCKET_LOCK_AMOUNT_IDENT: &str = "Bucket_lock_amount";
-
-#[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
-pub struct BucketLockAmountInput {
-    pub amount: Decimal,
-}
-
-pub type BucketLockAmountOutput = ();
-
-pub const BUCKET_UNLOCK_AMOUNT_IDENT: &str = "Bucket_unlock_amount";
-
-#[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
-pub struct BucketUnlockAmountInput {
-    pub amount: Decimal,
-}
-
-pub type BucketUnlockAmountOutput = ();
-
-pub const BUCKET_LOCK_NON_FUNGIBLES_IDENT: &str = "Bucket_lock_non_fungibles";
-
-#[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
-pub struct BucketLockNonFungiblesInput {
-    pub local_ids: BTreeSet<NonFungibleLocalId>,
-}
-
-pub type BucketLockNonFungiblesOutput = ();
-
-pub const BUCKET_UNLOCK_NON_FUNGIBLES_IDENT: &str = "Bucket_unlock_non_fungibles";
-
-#[derive(Debug, Clone, Eq, PartialEq, ScryptoSbor)]
-pub struct BucketUnlockNonFungiblesInput {
-    pub local_ids: BTreeSet<NonFungibleLocalId>,
-}
-
-pub type BucketUnlockNonFungiblesOutput = ();
 
 //========
 // Stub
