+++ radix-engine-interface/src/blueprints/resource/fungible/mod.rs
@@ -1,5 +1,7 @@
+mod fungible_bucket;
 mod fungible_resource_manager;
 mod fungible_vault;
 
+pub use fungible_bucket::*;
 pub use fungible_resource_manager::*;
 pub use fungible_vault::*;
