+++ radix-engine-interface/src/blueprints/resource/non_fungible/mod.rs
@@ -1,5 +1,7 @@
+mod non_fungible_bucket;
 mod non_fungible_resource_manager;
 mod non_fungible_vault;
 
+pub use non_fungible_bucket::*;
 pub use non_fungible_resource_manager::*;
 pub use non_fungible_vault::*;
